Add randomly selected workouts and double workout
